Comments on foreword
Here are comments following a read-through of the foreword in the form
of suggested edits, not in anticipation of merging this PR.

I suggest broadening the scope of the readership, to readers from
students and only refer to readers as students in the context of a
course, as introduced in a later paragraph.
I suggest removing some frequently repeated words, in particular "the
language"
A few suggestions of alternative words
High-level: I like the foreword. It is respectful of the source
material, and it is modest and to the point.