Addressing the issue of non-explained functions in §5.1
Following Martin's idea, we avoid relying on undefined functions that
establish a language that is mostly avoided in the next section.

For uniformity, Figure 5.3 uses a list to build a controller, but
starting on p. 485 we switch to a specific controller constructor,
and make a note in the text that this uses a tagged list internally.