Fixes #500
Line spacing fixed for
- 2.3
- 3.2 function square (Note: was already fixed)
- 3.4 rightmost RETURN (Note: was already fixed)
- 3.5 sum_of_squares (Note: was already fixed)
- 5.7

Excess whitespace trimmed for:
- 3.5, 3.20, 3.21

I have verified that: captions are typeset in small (same as
exercises) and code is the same size as in the exercises.

PDF_ONLY can now appear in FIGUREs so that I could add a
\flushleft to slide the code over to the left in 5.11.
Also, first comment is now on 1 line.

Reduced space between the two parts of 5.14.